Campus-Book-Trade/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-environ==0.4.54
requests==2.32.4
sqlparse==0.5.3
urllib3==2.4.0
